+++ plugins/woocommerce-admin/client/homescreen/activity-panel/index.js
@@ -10,7 +10,11 @@ import { getSetting } from '@woocommerce/wc-admin-settings';
  * Internal dependencies
  */
 import './style.scss';
-import { getOrderStatuses, getUnreadOrders } from './orders/utils';
+import {
+	getLowStockCount,
+	getOrderStatuses,
+	getUnreadOrders,
+} from './orders/utils';
 import { getAllPanels } from './panels';
 
 export const ActivityPanel = () => {
@@ -18,8 +22,13 @@ export const ActivityPanel = () => {
 		const totalOrderCount = getSetting( 'orderCount', 0 );
 		const orderStatuses = getOrderStatuses( select );
 		const countUnreadOrders = getUnreadOrders( select, orderStatuses );
+		const manageStock = getSetting( 'manageStock', 'no' );
+		const countLowStockProducts = getLowStockCount( select );
+
 		return getAllPanels( {
+			countLowStockProducts,
 			countUnreadOrders,
+			manageStock,
 			orderStatuses,
 			totalOrderCount,
 		} );