+++ plugins/woocommerce-admin/client/homescreen/activity-panel/panels.js
@@ -7,9 +7,12 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import OrdersPanel from './orders';
+import StockPanel from './stock';
 
 export function getAllPanels( {
+	countLowStockProducts,
 	countUnreadOrders,
+	manageStock,
 	orderStatuses,
 	totalOrderCount,
 } ) {
@@ -26,6 +29,16 @@ export function getAllPanels( {
 				/>
 			),
 			title: __( 'Orders', 'woocommerce-admin' ),
+		},
+		manageStock === 'yes' && {
+			className: 'woocommerce-homescreen-card',
+			count: countLowStockProducts,
+			id: 'stock-panel',
+			initialOpen: false,
+			panel: (
+				<StockPanel countLowStockProducts={ countLowStockProducts } />
+			),
+			title: __( 'Stock', 'woocommerce-admin' ),
 		},
 		// Add another panel row here
 	].filter( Boolean );