+++ plugins/woocommerce-admin/client/homescreen/activity-panel/test/index.js
@@ -33,6 +33,7 @@ jest.mock( '../panels', () => {
 // Mock the orders and order statuses.
 jest.mock( '../orders/utils', () => {
 	return {
+		getLowStockCount: jest.fn().mockImplementation( () => 0 ),
 		getUnreadOrders: jest.fn().mockImplementation( () => 100 ),
 		getOrderStatuses: jest.fn().mockImplementation( () => [ 'status' ] ),
 	};