+++ plugins/woocommerce-admin/packages/data/src/items/utils.js
@@ -7,6 +7,7 @@ import { appendTimestamp, getCurrentDates } from '@woocommerce/date';
  * Internal dependencies
  */
 import { STORE_NAME } from './constants';
+import { getResourceName } from '../utils';
 
 /**
  * Returns leaderboard data to render a leaderboard table.
@@ -91,4 +92,23 @@ export function searchItemsByString( select, endpoint, search ) {
 	} );
 
 	return { items, isRequesting, isError };
+}
+
+/**
+ * Generate a resource name for item totals count.
+ *
+ * It omits query parameters from the identifier that don't affect
+ * totals values like pagination and response field filtering.
+ *
+ * @param {string} itemType Item type for totals count.
+ * @param {Object} query Query for item totals count.
+ * @return {string} Resource name for item totals.
+ */
+export function getTotalCountResourceName( itemType, query ) {
+	// Disable eslint rule because we're using this spread to omit properties
+	// that don't affect item totals count results.
+	// eslint-disable-next-line no-unused-vars, camelcase
+	const { _fields, page, per_page, ...totalsQuery } = query;
+
+	return getResourceName( 'total-' + itemType, totalsQuery );
 }