+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+-   Change styling of `<ProductImage />`
+-   Add new `<Accordion>` component.
+
 # 5.1.2
 
 -   Update dependencies.